Reorganize models/layers/factories folder
