+++ src/app/modules/core/models/layers/object3d-factory.ts
