+++ src/app/modules/core/models/layers/text/text-offsets.ts
