+++ src/app/modules/core/models/layers/text/top-right-text.spec.ts
@@ -1,4 +1,4 @@
-import { TopRightText } from '#core/models/layers/factories/text/top-right-text';
+import { TopRightText } from '#core/models/layers/text/top-right-text';
 
 
 describe('TopRightText', () => {
