+++ src/app/modules/core/services/factories/aggregate-layer-factory.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { AggregateLayer } from '#core/models/layers/aggregate-layer';
-import { LayerFactory } from '#core/models/layers/factories/layer-factory';
+import { LayerFactory } from '#core/models/layers/layer-factory';
 import { Layer } from '#core/models/layers/layer';
 
 /**
