+++ src/app/modules/layer-constellations/services/factories/lines-layer-factory.service.spec.ts
@@ -1,7 +1,7 @@
 import { TestBed } from '@angular/core/testing';
 import { AxialCurvesFactoryService } from '#core/services/factories/axial-curves-factory.service';
 import { LinesLayerFactoryService } from '#layer-constellations/services/factories/lines-layer-factory.service';
-import { LayerFactory } from '#core/models/layers/factories/layer-factory';
+import { LayerFactory } from '#core/models/layers/layer-factory';
 
 
 describe('LinesLayerFactoryService', () => {
