+++ src/app/modules/layer-constellations/services/factories/lines-layer-factory.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@angular/core';
 import { Layer } from '#core/models/layers/layer';
 import { ConstellationLines } from '#layer-constellations/models/constellation-lines';
 import { AxialCurvesFactoryService } from '#core/services/factories/axial-curves-factory.service';
-import { LayerFactory } from '#core/models/layers/factories/layer-factory';
+import { LayerFactory } from '#core/models/layers/layer-factory';
 import { SupportedLayers } from '#core/models/layers/supported-layers';
 
 /**
