+++ src/app/modules/layer-constellations/services/factories/names-layer-factory.service.spec.ts
@@ -1,6 +1,6 @@
 import { TestBed } from '@angular/core/testing';
 import { NamesLayerFactoryService } from '#layer-constellations/services/factories/names-layer-factory.service';
-import { LayerFactory } from '#core/models/layers/factories/layer-factory';
+import { LayerFactory } from '#core/models/layers/layer-factory';
 
 
 describe('NamesLayerFactoryService', () => {
