+++ src/app/modules/layer-constellations/services/factories/names-layer-factory.service.ts
@@ -2,8 +2,8 @@ import { Injectable } from '@angular/core';
 import { ConstellationMeta } from '#core/models/layers/constellation-meta';
 import { Layer } from '#core/models/layers/layer';
 import { ConstellationNames } from '#layer-constellations/models/constellation-names';
-import { LayerFactory } from '#core/models/layers/factories/layer-factory';
-import { TextOffsetPolicies } from '#core/models/layers/factories/text/text-offsets-policies';
+import { LayerFactory } from '#core/models/layers/layer-factory';
+import { TextOffsetPolicies } from '#core/models/layers/text/text-offsets-policies';
 import { RenderableText } from '#core/models/layers/renderable-text';
 import { WorldConstants } from '#core/models/world-constants';
 import { toVector3 } from '#core/utils/vector-utils';
